requirements.txt
# Python Dependencies for ML Data Pipeline

# Kafka Producer/Consumer
confluent-kafka==2.3.0

# Spark (for streaming processing)
pyspark==3.5.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Configuration management
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
